+++ CHANGE_LOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+
+- Uptake [@microsoft/omnichannel-chat-components@1.0.6](https://www.npmjs.com/package/@microsoft/omnichannel-chat-components/v/1.0.6)
+
+### Fixed
+
+- Fixed an issue where video card cannot be rendered
+
 ## [1.2.2] - 2023-8-16
 
 ### Changed